Test some edge examples from Pyro (#131)
